pyproject.toml
[project]
name = "mcp-sse-client"
version = "0.1.0"
description = "SSE-based MCP Client for connecting to MCP servers"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.0.0",
    "mcp[cli]>=1.2.1",
    "python-dotenv>=1.0.1",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langgraph>=0.2.0",
    "langgraph-cloud>=0.1.0",
    "httpx>=0.28.1",
    "pydantic>=2.0.0",
]